Refactor and add update method
FS class refactored to now accept a validator module and a formatting module. Removed file instance method and references to it; replaced with local `file` variable when needed. Added headers_list parameter to create file, and refactored so implementation now knows to add formatted headers to the new file on creation. Added `update_file()` method to allow an Admin to add new data to a pre-existing file. Added private `__get_column_count` method to calculate the maximum amount of columns needed when writing data to the pre-existing file.